fix: Corregir incompatibilidad de tipos en Foreign Keys - Cambiar orders.id de int(11) a bigint(20) para compatibilidad con users.id - Cambiar order_items.order_id de int(11) a bigint(20) para compatibilidad con orders.id - Cambiar order_status_history.order_id de int(11) a bigint(20) para compatibilidad con orders.id - Esto resuelve el error 'Referencing column user_id and referenced column id are incompatible'

+++ backend/src/config/database.js
@@ -145,7 +145,7 @@ const createAllTables = async (connection) => {
     // 8. Crear tabla de órdenes
     await connection.query(`
       CREATE TABLE IF NOT EXISTS orders (
-        id int(11) NOT NULL AUTO_INCREMENT,
+        id bigint(20) NOT NULL AUTO_INCREMENT,
         order_number varchar(50) NOT NULL,
         customer_type enum('registered','guest') NOT NULL,
         user_id bigint(20) DEFAULT NULL,
@@ -184,7 +184,7 @@ const createAllTables = async (connection) => {
     await connection.query(`
       CREATE TABLE IF NOT EXISTS order_items (
         id int(11) NOT NULL AUTO_INCREMENT,
-        order_id int(11) NOT NULL,
+        order_id bigint(20) NOT NULL,
         product_id bigint(20) NOT NULL,
         product_name varchar(255) NOT NULL,
         product_price decimal(10,2) NOT NULL,
@@ -353,7 +353,7 @@ const createAllTables = async (connection) => {
     await connection.query(`
       CREATE TABLE IF NOT EXISTS order_status_history (
         id int(11) NOT NULL AUTO_INCREMENT,
-        order_id int(11) NOT NULL,
+        order_id bigint(20) NOT NULL,
         previous_status enum('pending','confirmed','preparing','ready','delivered','cancelled') DEFAULT NULL,
         new_status enum('pending','confirmed','preparing','ready','delivered','cancelled') NOT NULL,
         changed_by enum('system','admin','customer') NOT NULL,
@@ -524,7 +524,7 @@ const createSingleTable = async (connection, tableName) => {
       case 'orders':
         await connection.query(`
           CREATE TABLE IF NOT EXISTS orders (
-            id int(11) NOT NULL AUTO_INCREMENT,
+            id bigint(20) NOT NULL AUTO_INCREMENT,
             order_number varchar(50) NOT NULL,
             customer_type enum('registered','guest') NOT NULL,
             user_id bigint(20) DEFAULT NULL,
@@ -564,7 +564,7 @@ const createSingleTable = async (connection, tableName) => {
         await connection.query(`
           CREATE TABLE IF NOT EXISTS order_items (
             id int(11) NOT NULL AUTO_INCREMENT,
-            order_id int(11) NOT NULL,
+            order_id bigint(20) NOT NULL,
             product_id bigint(20) NOT NULL,
             product_name varchar(255) NOT NULL,
             product_price decimal(10,2) NOT NULL,
@@ -741,7 +741,7 @@ const createSingleTable = async (connection, tableName) => {
         await connection.query(`
           CREATE TABLE IF NOT EXISTS order_status_history (
             id int(11) NOT NULL AUTO_INCREMENT,
-            order_id int(11) NOT NULL,
+            order_id bigint(20) NOT NULL,
             previous_status enum('pending','confirmed','preparing','ready','delivered','cancelled') DEFAULT NULL,
             new_status enum('pending','confirmed','preparing','ready','delivered','cancelled') NOT NULL,
             changed_by enum('system','admin','customer') NOT NULL,
